Moved is_null and is_not_null
